ML_Basics/server_failure_rca/requirements.txt
pandas
scikit-learn
matplotlib
pyyaml
joblib
fastapi
uvicorn[standard]
